Changed order of debug statements and base class __init__ calls.
Added an objectList to the scene object, so that it can keep track of
objects within a scene, and therefore call their render() methods, which is
now implemented in Scene.render().  Each object should call scene.add(self)
at __init__ so that it can be added to the object list in the scene.  This
is different for Image() objects, since they can be used without being added
to the scene, as in the save() method for example.


git-svn-id: https://pyvisi.svn.sourceforge.net/svnroot/pyvisi@458 4460aaa7-681d-0410-9b2c-fbf292dd6b46