Changed order of debug statements and base class __init__ calls.
Added an objectList so that scene objects can keep a record of the objects
that are within a scene and so call the individual render() methods when
scene.render() is called.  Added the ability to add objects to the object
list in scene.add().  All objects that need to be in the object list should
call scene.add(self) at __init__.


git-svn-id: https://pyvisi.svn.sourceforge.net/svnroot/pyvisi@460 4460aaa7-681d-0410-9b2c-fbf292dd6b46